Register flow. Extraction of form elements for Register/Login
